Split ORM from Logic
