+++ lib/sidekiq/activerecord.rb
@@ -7,10 +7,19 @@ require 'sidekiq/active_record/version'
 
 
 module Sidekiq
+
+  module Orm
+    extend ActiveSupport::Autoload
+
+    autoload :TaskWorker
+    autoload :ManagerWorker
+  end
+
   module ActiveRecord
     extend ActiveSupport::Autoload
 
     autoload :TaskWorker
     autoload :ManagerWorker
   end
+
 end
